refactor code to follow coding standards
