+++ src/main/java/duke/Duke.java
@@ -1,11 +1,11 @@
 package duke;
 
-import duke.customClass.Command;
-import duke.customClass.Ui;
-import duke.customClass.Storage;
-import duke.customClass.TaskList;
-import duke.customClass.Parser;
-import duke.customClass.CommandRouter;
+import duke.processintructions.Command;
+import duke.ui.Ui;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.processintructions.Parser;
+import duke.processintructions.CommandRouter;
 
 /**
  * Duke is a todo list with features built incrementally.
@@ -38,6 +38,9 @@ public class Duke {
 
     /**
      * This method call each of the respective classes to execute a command that the user input
+     *
+     * @param input Input string typed by the user.
+     * @return Output string generated by Duke in response.
      */
     public String getResponse(String input) {
         CommandRouter commandRouter = new CommandRouter();
