+++ src/main/java/duke/processintructions/Command.java
@@ -1,4 +1,4 @@
-package duke.customClass;
+package duke.processintructions;
 
 public enum Command {
     TODO,
