+++ src/main/java/duke/processintructions/Find.java
@@ -1,4 +1,4 @@
-package duke.customClass;
+package duke.processintructions;
 
 import duke.tasks.Task;
