+++ src/main/java/duke/tasks/Event.java
@@ -4,6 +4,10 @@ import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
+/**
+ * Event is a type of Task which has a description and date.
+ * It also maintains a state of isDone.
+ */
 public class Event extends Task {
     protected String at;
 
@@ -19,8 +23,9 @@ public class Event extends Task {
     }
 
     /**
-     * returns string representation of Event when saving locally.
-     * @return string of Event when saving.
+     * Returns string representation of Event when saving locally.
+     *
+     * @return String of Event when saving.
      */
     public String saveString() {
         return isDone ? "E --- 1 --- " + description + " --- " + at : "E --- 0 --- " + description + " --- " + at;
@@ -28,8 +33,9 @@ public class Event extends Task {
 
     /**
      * Converts user input for date in yyyy-mm-dd format into MMM d yyyy format.
-     * @param input a string representation of a date in the format yyyy-mm-dd.
-     * @return string representation of date in MMM d yyyy format.
+     *
+     * @param input A string representation of a date in the format yyyy-mm-dd.
+     * @return String representation of date in MMM d yyyy format.
      */
     public static String convertToDate(String input) {
         try {
@@ -41,8 +47,9 @@ public class Event extends Task {
     }
 
     /**
-     * returns string representation of Event.
-     * @return string of Event.
+     * Returns string representation of Event.
+     *
+     * @return String of Event.
      */
     @Override
     public String toString() {
